status code check in controllers, response creator, new checks in services and refactoring

+++ Recipes.BLL/Services/CookingStepService.cs
@@ -1,11 +1,10 @@
 using AutoMapper;
-using Recipes.BLL.Interfaces;
-using Recipes.DAL.Interfaces;
+using Recipes.BLL.Helpers;
+using Recipes.BLL.Services.Interfaces;
+using Recipes.DAL.Infrastructure.Interfaces;
 using Recipes.Data.DataTransferObjects;
-using Recipes.Data.Enums;
-using Recipes.Data.Interfaces;
 using Recipes.Data.Models;
-using Recipes.Data.Responses;
+using Recipes.Data.Responses.Interfaces;
 
 namespace Recipes.BLL.Services;
 
@@ -13,11 +12,13 @@ public class CookingStepService : ICookingStepService
 {
     private readonly IUnitOfWork _unitOfWork;
     private readonly IMapper _mapper;
+    private readonly ResponseCreator _responseCreator;
 
     public CookingStepService(IUnitOfWork unitOfWork, IMapper mapper)
     {
         _unitOfWork = unitOfWork;
         _mapper = mapper;
+        _responseCreator = new ResponseCreator();
     }
 
     public Task<IBaseResponse<CookingStepDto>> GetById(Guid id)
@@ -25,24 +26,22 @@ public class CookingStepService : ICookingStepService
         throw new NotImplementedException();
     }
 
-    public async Task<IBaseResponse<IEnumerable<CookingStepDto>>> Get()
+    public async Task<IBaseResponse<List<CookingStepDto>>> Get()
     {
         try
         {
             var models = await _unitOfWork.CookingStepRepository.GetAsync();
 
             if (models.Count is 0)
-            {
-                return BaseResponse<CookingStepDto>.CreateBaseResponse<IEnumerable<CookingStepDto>>("0 objects found", StatusCode.NotFound);
-            }
+                return _responseCreator.CreateBaseNotFound<List<CookingStepDto>>("No cooking steps found.");
 
             var dtoList = models.Select(model => _mapper.Map<CookingStepDto>(model)).ToList();
 
-            return BaseResponse<CookingStepDto>.CreateBaseResponse<IEnumerable<CookingStepDto>>("Success!", StatusCode.Ok, dtoList, dtoList.Count);
+            return _responseCreator.CreateBaseOk(dtoList, dtoList.Count);
         }
         catch(Exception e) 
         {
-            return BaseResponse<CookingStepDto>.CreateBaseResponse<IEnumerable<CookingStepDto>>(e.Message, StatusCode.InternalServerError);
+            return _responseCreator.CreateBaseServerError<List<CookingStepDto>>(e.Message);
         }
     }
 
@@ -51,19 +50,19 @@ public class CookingStepService : ICookingStepService
         try
         {
             if (modelDto is null) 
-                return BaseResponse<CookingStepDto>.CreateBaseResponse<string>("Objet can`t be empty...", StatusCode.BadRequest);
+                return _responseCreator.CreateBaseBadRequest<string>("Inserted cooking step is empty.");
             
             modelDto.Id = Guid.NewGuid();
                 
             await _unitOfWork.CookingStepRepository.InsertAsync(_mapper.Map<CookingStep>(modelDto));
             await _unitOfWork.SaveChangesAsync();
 
-            return BaseResponse<CookingStepDto>.CreateBaseResponse<string>("Object inserted!", StatusCode.Ok, resultsCount: 1);
+            return _responseCreator.CreateBaseOk($"Cooking step added.", 1);
 
         }
         catch (Exception e)
         {
-            return BaseResponse<CookingStepDto>.CreateBaseResponse<string>(e.Message, StatusCode.InternalServerError);
+            return _responseCreator.CreateBaseServerError<string>(e.Message);
         }
     }
 
@@ -71,16 +70,17 @@ public class CookingStepService : ICookingStepService
     {
         try
         {
+            if (id == Guid.Empty)
+                return _responseCreator.CreateBaseBadRequest<string>("Id is empty.");
+            
             await _unitOfWork.CookingStepRepository.DeleteAsync(id);
             await _unitOfWork.SaveChangesAsync();
 
-            return BaseResponse<CookingStepDto>.CreateBaseResponse<string>("Object deleted!", StatusCode.Ok, resultsCount: 1);
+            return _responseCreator.CreateBaseOk("Cooking step deleted.", 1);
         }
         catch (Exception e)
         {
-            return BaseResponse<CookingStepDto>.CreateBaseResponse<string>($"{e.Message} or object not found", StatusCode.InternalServerError);
+            return _responseCreator.CreateBaseServerError<string>(e.Message);
         }
     }
-    
-   
 }
